C9 Pro MM source SM-C900F kernel merge samsung build tag C900FDDU1AQC3 merged on CAF tag LA.BR.1.3.6.c2_rb2.3

+++ drivers/iommu/msm_iommu_sec.c
@@ -82,7 +82,7 @@ struct msm_scm_paddr_list {
 struct msm_scm_mapping_info {
 	unsigned int id;
 	unsigned int ctx_id;
-	unsigned int va;
+	unsigned long va;
 	unsigned int size;
 };
